Use large range on true "dev" dependencies (#114)
* Remove unnecessary dev dependencies

* Use large range on true "dev" dependencies

The exact versions of these dependencies are not important for building, so give them large ranges in order to quiet down greenkeeper.

I did not split them between dependencies and devDependencies because a package's dependencies are installed whenever the package is installed as a dependency. Installing all of these would be obnoxious and unnecessary.