feat: allow MenuBar to be hidden
The MenuBar is quite specific to the canonical version of the online Scratch
editor [1]. In RaspberryPiFoundation projects we want to make use of the
scratch-gui package without displaying the MenuBar.

The MenuBar can be hidden by passing in the `menuBarHidden` prop.

Since the height of the menu bar is hard-coded as a CSS variable, I've
had to override the height calculation for the body wrapper when the
menu bar is not visible. Ideally the height of the menu bar wouldn't be
hard-coded and this wouldn't be necessary.

We were unable to find an existing unit test for the Gui component and so we
have not added any test coverage around the new `menuBarHidden` prop.

[1]: https://scratch.mit.edu/projects/editor

Co-authored-by: James Mead <james.mead@gofreerange.com>