test: adjust project execution tests for tap >=18
Calling `t.plan()` has more significant consequences now in terms of
test teardown. In particular, calling `t.plan(0)` is a bit hard to deal
with correctly. Adding 1 to the plan count reported by the project, and
consistently adding exactly one test after the project finishes, makes
the behavior pretty much the same as it was with tap 16. I also used
`async` and `await` to make some of this code easier to read.