Simplifies update process
Streamlines the update mechanism by downloading and executing the installation script directly from the repository.

This approach eliminates the need for managing and loading container images within the update process, and relies on the install script to do it, which simplifies maintenance.

It also adds a check for curl to ensure that the update mechanism is working.