Adds AlbiLAB IP address configuration wizard
Introduces a wizard to guide users in setting up the IP address for AlbiLAB devices.

Adds a modal that prompts the user for the IP address upon initial load, which is then stored in local storage. Also includes a menu item to change the IP later.

The page reloads after saving to apply the new IP across all AlbiLAB components.